source/requirements.txt
## ffmpeg or avconv required
azure-storage-blob  
speechbrain
torchaudio
speechbrain
transformers
pydub
azurebatchload